AnalogSensorClient/wrapper: update to the new policy for parameters, move client to a folder inside modules
