+++ src/libYARP_dev/src/modules/AnalogWrapper/AnalogWrapper.cpp
@@ -5,6 +5,7 @@
  * CopyPolicy: Released under the terms of the LGPLv2.1 or later, see LGPL.TXT
  */
 
+#include <sstream>
 #include <iostream>
 #include "AnalogWrapper.h"
 #include <yarp/dev/ControlBoardInterfaces.h>
@@ -195,8 +196,7 @@ void AnalogWrapper::setHandlers()
     for(unsigned int i=0;i<analogPorts.size(); i++)
     {
         std::string rpcPortName = analogPorts[i].port_name;
-        rpcPortName += "/rpc:i";
-//        analogPorts[i].port_name += ":o";
+        rpcPortName += "/rpc";
         AnalogServerHandler* ash = new AnalogServerHandler(rpcPortName.c_str());
         handlers.push_back(ash);
     }
@@ -290,31 +290,20 @@ bool AnalogWrapper::threadInit()
     return true;
 }
 
-void AnalogWrapper::setId(const std::string &i)
+void AnalogWrapper::setId(const std::string &id)
 {
-    id=i;
+    sensorId=id;
 }
 
 std::string AnalogWrapper::getId()
 {
-    return id;
+    return sensorId;
 }
 
-Bottle AnalogWrapper::getOptions()
+bool AnalogWrapper::checkForDeprecatedParams(yarp::os::Searchable &params)
 {
-    Bottle options;
-    options.addString("robotName, mandatory");
-    options.addString("deviceId, mandatory");
-    options.addString("period");
-    return options;
-}
-
-bool AnalogWrapper::open(yarp::os::Searchable &config)
-{
-    Property params;
-    params.fromString(config.toString().c_str());
-
-    // Verify minimum set of parameters required
+//    check for deprecated params
+    bool correct = true;
     if(!params.check("robotName", "name of the robot.") )
     {
         yError("AnalogWrapper: missing robotName, check your configuration file");
@@ -323,39 +312,79 @@ bool AnalogWrapper::open(yarp::os::Searchable &config)
 
     if(params.check("deviceId"))
     {
-      string tmp(params.find("deviceId").asString());
+        string tmp(params.find("deviceId").asString());
       //yDebug() << "AnalogWrapper Debug" << tmp;
-      setId(tmp);
+        setId(tmp);
     }
     else
     {
         yError("AnalogWrapper: no device Id found");
+        cerr << "ERROR: AnalogServer missing DEPRECATED parameter 'deviceId', check your configuration file!! Quitting\n";
+    }
+
+    if(!correct)
         return false;
+
+    std::cerr << "************************************************************************************************";
+    std::cerr << " WARNING: using deprecated configuration, parameters for creating port names should be:";
+    std::cerr << "          name:         full name of the port, like /robotName/deviceId/sensorType:o";
+    std::cerr << "          period:       refresh period of the broadcasted values in ms (optional, default 20ms)";
+    std::cerr << "************************************************************************************************";
+
+    // Create the list of ports
+    std::string robotName = params.find("robotName").asString().c_str();
+    streamingPortName ="/";
+    streamingPortName += robotName;
+    streamingPortName += "/" + this->sensorId + "/analog:o";
+}
+
+bool AnalogWrapper::open(yarp::os::Searchable &config)
+{
+    Property params;
+    params.fromString(config.toString().c_str());
+
+    std::cout << "\nAS\n\n" << params.toString() << "\n\n" << std::endl;
+    bool correct=true;
+
+    // Verify minimum set of parameters required
+    streamingPortName.clear();
+    rpcPortName.clear();
+
+    if (!config.check("name"))
+    {
+        correct = checkForDeprecatedParams(config);
+        if(!correct)
+        {
+            std::cerr << "AnalogServer: missing 'name' parameter. Check you configuration file. Should look like:";
+            std::cerr << " name:         full name of the port, like /robotName/deviceId/sensorType:o";
+        }
+    }
+    else
+    {
+        streamingPortName  = config.find("name").asString().c_str();
+        setId("AnalogServer");
     }
 
     if (params.check("period"))
     {
-        _rate=params.find("period").asInt();
+        _rate = params.find("period").asInt();
     }
     else
     {
-        _rate=20;
+        _rate = DEFAULT_THREAD_PERIOD;
          yWarning("AnalogWrapper: part %s using default period %d", id.c_str() , _rate);
     }
 
-    // Create the list of ports
-    std::string robotName=params.find("robotName").asString().c_str();
-    std::string root_name;
-    root_name+="/";
-    root_name+=robotName;
-    root_name+= "/" + this->id + "/analog:o";
+    if(!correct)
+        return false;
+
 
     // port names are optional, do not check for correctness.
     if(!params.check("ports"))
     {
      // if there is no "ports" section open only 1 port and use name as is.
-        createPort((root_name ).c_str(), _rate );
-        yDebug() << "opening port " << root_name.c_str();
+        createPort((streamingPortName ).c_str(), _rate );
+        cout << "opening port " << streamingPortName.c_str();
     }
     else
     {
@@ -402,7 +431,7 @@ bool AnalogWrapper::open(yarp::os::Searchable &config)
             tmpPorts[k].length = portChannels;
             tmpPorts[k].offset = wBase;
             yDebug() << "opening port " << ports->get(k).asString().c_str();
-            tmpPorts[k].port_name = root_name+ "/" + string(ports->get(k).asString().c_str());
+            tmpPorts[k].port_name = streamingPortName+ "/" + string(ports->get(k).asString().c_str());
 
             createPorts(tmpPorts, _rate);
             sumOfChannels+=portChannels;
@@ -414,6 +443,8 @@ bool AnalogWrapper::open(yarp::os::Searchable &config)
             return false;
         }
     }
+
+    threadInit();  // debug only TBR
     return true;
 }
